just push harder

+++ wiselib.testing/external_interface/pc/pc_os_model.h
@@ -60,8 +60,8 @@ namespace wiselib {
 			typedef PCOsModel AppMainParameter;
 			typedef PCOsModel Os;
 			
-			//typedef uint32_t size_t;
-			typedef unsigned long size_t;
+			typedef uint32_t size_t;
+//			typedef unsigned long size_t;
 			typedef uint8_t block_data_t;
 			
 			typedef PCClockModel<PCOsModel> Clock;
